+++ modules/node/src/channel/channel.service.ts
@@ -1,9 +1,9 @@
 import { CreateChannelMessage } from "@counterfactual/node";
 import { Node as NodeTypes } from "@counterfactual/types";
-import { Injectable, OnModuleInit } from "@nestjs/common";
+import { Injectable } from "@nestjs/common";
 import { RpcException } from "@nestjs/microservices";
 import { AddressZero } from "ethers/constants";
-import { BigNumber, bigNumberify } from "ethers/utils";
+import { BigNumber } from "ethers/utils";
 
 import { NodeService } from "../node/node.service";
 import { PaymentProfile } from "../paymentProfile/paymentProfile.entity";
@@ -15,12 +15,16 @@ import { ChannelRepository } from "./channel.repository";
 const logger = new CLogger("ChannelService");
 
 @Injectable()
-export class ChannelService implements OnModuleInit {
+export class ChannelService {
   constructor(
     private readonly nodeService: NodeService,
     private readonly channelRepository: ChannelRepository,
   ) {}
 
+  /**
+   * Starts create channel process within CF node
+   * @param counterpartyPublicIdentifier
+   */
   async create(counterpartyPublicIdentifier: string): Promise<NodeTypes.CreateChannelResult> {
     logger.log(`Creating channel for ${counterpartyPublicIdentifier}`);
     const existing = await this.channelRepository.findByUserPublicIdentifier(
@@ -86,35 +90,34 @@ export class ChannelService implements OnModuleInit {
     return await this.channelRepository.addPaymentProfileToChannel(userPubId, profile);
   }
 
-  onModuleInit(): void {
-    this.nodeService.registerCfNodeListener(
-      NodeTypes.EventName.CREATE_CHANNEL,
-      async (creationData: CreateChannelMessage) => {
-        const existing = await this.channelRepository.findByMultisigAddress(
-          creationData.data.multisigAddress,
-        );
-        if (existing) {
-          if (
-            !creationData.data.owners.includes(existing.nodePublicIdentifier) ||
-            !creationData.data.owners.includes(existing.userPublicIdentifier)
-          ) {
-            throw new Error(
-              `Channel has already been created with different owners! ${JSON.stringify(
-                existing,
-              )}. Event data: ${creationData}`,
-            );
-          }
-          logger.log(`Channel already exists in database`);
-        }
-        logger.log(`Creating new channel from data ${JSON.stringify(creationData)}`);
-        const channel = new Channel();
-        channel.userPublicIdentifier = creationData.data.counterpartyXpub;
-        channel.nodePublicIdentifier = this.nodeService.cfNode.publicIdentifier;
-        channel.multisigAddress = creationData.data.multisigAddress;
-        channel.available = true;
-        await this.channelRepository.save(channel);
-      },
-      logger.cxt,
+  /**
+   * Creates a channel in the database with data from CF node event CREATE_CHANNEL
+   * and marks it as available
+   * @param creationData event data
+   */
+  async makeAvailable(creationData: CreateChannelMessage): Promise<void> {
+    const existing = await this.channelRepository.findByMultisigAddress(
+      creationData.data.multisigAddress,
     );
+    if (existing) {
+      if (
+        !creationData.data.owners.includes(existing.nodePublicIdentifier) ||
+        !creationData.data.owners.includes(existing.userPublicIdentifier)
+      ) {
+        throw new Error(
+          `Channel has already been created with different owners! ${JSON.stringify(
+            existing,
+          )}. Event data: ${creationData}`,
+        );
+      }
+      logger.log(`Channel already exists in database`);
+    }
+    logger.log(`Creating new channel from data ${JSON.stringify(creationData)}`);
+    const channel = new Channel();
+    channel.userPublicIdentifier = creationData.data.counterpartyXpub;
+    channel.nodePublicIdentifier = this.nodeService.cfNode.publicIdentifier;
+    channel.multisigAddress = creationData.data.multisigAddress;
+    channel.available = true;
+    await this.channelRepository.save(channel);
   }
 }
