+++ modules/node/src/listener/listener.module.ts
@@ -1,0 +1,9 @@
+import { Module } from "@nestjs/common";
+
+@Module({
+  controllers: [],
+  exports: [],
+  imports: [],
+  providers: [],
+})
+export class ListenerModule {}
